requirements.txt
fastapi==0.121.0
httpx==0.28.1
pydantic==2.12.4
python-dotenv==1.2.1
uvicorn==0.38.0
